requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
anyio==4.7.0
asgiref==3.8.1
asttokens==2.4.1
asyncpg==0.30.0
atomicwrites==1.4.1
attrs==24.3.0
autopep8==2.3.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==25.1.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
diff-match-patch==20241021
distlib==0.3.9
dj-database-url==2.3.0
Django==4.2.18
django-admin-rangefilter==0.13.2
django-allauth==65.3.0
django-anymail==12.0
django-constance==4.1.3
django-cors-headers==4.6.0
django-db-geventpool==4.0.7
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==25.1
django-import-export==4.3.5
django-mailgun-provider==0.2.3
django-redis==5.4.0
django-rest-passwordreset==1.5.0
django-silk==5.3.2
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
dulwich==0.22.7
executing==2.1.0
factory_boy==3.3.3
Faker==36.1.1
fastapi==0.68.2
fastjsonschema==2.21.1
filelock==3.17.0
flake8==7.1.2
fonttools==4.56.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.45.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gprof2dot==2024.6.6
gunicorn==23.0.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
identify==2.6.7
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.29.5
ipython==8.27.0
isort==6.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.8
kombu==5.5.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
pipdeptree==2.24.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
pre_commit==4.1.0
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==1.10.19
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.1.4
pyproject_hooks==1.2.0
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
pywin32==308; platform_system == "Windows"
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.1
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
shellingham==1.5.4
simplejwt==2.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.14.2
tablib==3.8.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.1.15.22
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.15.0
vine==5.1.0
virtualenv==20.29.1
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.7.0
XlsxWriter==3.2.2
yarl==1.18.3
